Refactoring direcroty arch
