+++ src/project/ProjectDetailPage.tsx
@@ -1,48 +1,33 @@
 import React from "react";
-import { useRouter } from "next/router";
-import Link from "next/link";
 
-import { styled } from "../../stitches.config";
-import { Project } from "../type";
+import { styled } from "../stitches.config";
 
-import { ProjectTopMessages } from "./ProjectTopMessages";
+import { Project } from "./type";
 import { CopyProjectLink } from "./CopyProjectLink";
 import { MusicEdit } from "./MusicEdit";
+import { SomeTextMessages } from "./SomeTextMessages";
+import { SomeImageMessages } from "./SomeImageMessage";
 
 type Props = {
   project: Project;
 };
 
 export const ProjectDetailPage: React.FC<Props> = (props) => {
-  const router = useRouter();
-  const { project_id } = router.query;
-
-  const messageData = {
-    textMessageData: props.project.text_messages,
-    imageMessageData: props.project.image_messages,
-  };
-
   return (
     <Container>
       <ReceiverName>{props.project.receiver_name}さんへの色紙</ReceiverName>
+
       <ButtonWrapper>
         <PreviewButton>プレビューを表示</PreviewButton>
         <CompleteButton>完成する</CompleteButton>
       </ButtonWrapper>
+
       <CopyProjectLink />
+
       <MusicEdit></MusicEdit>
-      {/* <SpotifyMusicSelect {...SampleMusic}></SpotifyMusicSelect> */}
-      <ProjectTopMessages {...messageData} />
-      <div>
-        <NewMessageLink
-          href={{
-            pathname: "/projects/[project_id]/new_message",
-            query: { project_id: project_id },
-          }}
-        >
-          思いをとどける
-        </NewMessageLink>
-      </div>
+
+      <SomeTextMessages textMessageData={props.project.text_messages} />
+      <SomeImageMessages imageMessageData={props.project.image_messages} />
     </Container>
   );
 };
@@ -89,25 +74,4 @@ const ReceiverName = styled("h2", {
   fontSize: "24px",
   fontWeight: "700",
   textAlign: "left",
-});
-
-const NewMessageLink = styled(Link, {
-  color: "#FFFFFF",
-  background: "$yellow900",
-  fontWeight: "700",
-  fontSize: "20px",
-  width: "80%",
-  maxWidth: "100%",
-  height: "61px",
-  borderRadius: "44px",
-  border: "none",
-  margin: "30px auto",
-  textDecoration: "none",
-  display: "flex",
-  justifyContent: "center",
-  alignItems: "center",
-
-  position: "fixed",
-  bottom: 0,
-  left: "10%",
 });
