+++ src/project/SomeImageMessage.tsx
@@ -3,48 +3,23 @@ import { useRouter } from "next/router";
 import React from "react";
 import { IoChevronForwardOutline } from "react-icons/io5";
 
-import { styled } from "../../stitches.config";
-import { Message, SenderName, TextMessageContainer, TextMessageUnit } from "../TextMessagesPage";
-import { ImageMessage, TextMessage } from "../type";
+import { styled } from "../stitches.config";
+
+import { ImageMessage } from "./type";
 
 type Props = {
-  textMessageData: {
-    count: number;
-    items: TextMessage[];
-  };
   imageMessageData: {
     count: number;
     items: ImageMessage[];
   };
 };
 
-export const ProjectTopMessages: React.FC<Props> = (props) => {
+export const SomeImageMessages: React.FC<Props> = (props) => {
   const router = useRouter();
   const project_id = router.query.project_id;
 
   return (
     <Container>
-      <ItemsCaption>
-        <CaptionTitle>{props.textMessageData.count}件のメッセージ</CaptionTitle>
-        <BrowseAllLink href={`/projects/${project_id}/text_messages`}>
-          <LinkText>すべて見る</LinkText>
-          <LinkIcon>
-            <IoChevronForwardOutline />
-          </LinkIcon>
-        </BrowseAllLink>
-      </ItemsCaption>
-
-      <TextMessageContainer>
-        {props.textMessageData.items.map((messageData, i) => {
-          return (
-            <TextMessageUnit key={i}>
-              <Message>{messageData.text}</Message>
-              <SenderName>{messageData.sender_name}</SenderName>
-            </TextMessageUnit>
-          );
-        })}
-      </TextMessageContainer>
-
       <ItemsCaption>
         <CaptionTitle>{props.imageMessageData.count}枚の写真</CaptionTitle>
         <BrowseAllLink href={`/projects/${project_id}/image_messages`}>
