+++ src/project/newMessage/NewMessagePage.tsx
@@ -2,7 +2,7 @@ import Link from "next/link";
 import { useRouter } from "next/router";
 import React, { useState } from "react";
 
-import { styled } from "../stitches.config";
+import { styled } from "../../stitches.config";
 
 import { ImageMessageEdit } from "./ImageMessageEdit";
 import { SenderNameEdit } from "./SenderNameEdit";
