+++ src/project/newMessage/SenderNameEdit.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { styled } from "../stitches.config";
+import { styled } from "../../stitches.config";
 
 import { SectionTitle } from "./NewMessagePage";
